+++ salmon/data/prototype/config/boot.py
@@ -1,7 +1,7 @@
 import logging.config
 
 from salmon.routing import Router
-from salmon.server import Relay, SMTPReceiver
+from salmon.server import Relay, LMTPReceiver
 from salmon import queue
 
 from config import settings
@@ -14,7 +14,7 @@ settings.relay = Relay(host=settings.relay_config['host'],
                        port=settings.relay_config['port'], debug=1)
 
 # where to listen for incoming messages
-settings.receiver = SMTPReceiver(settings.receiver_config['host'],
+settings.receiver = LMTPReceiver(settings.receiver_config['host'],
                                  settings.receiver_config['port'])
 
 Router.defaults(**settings.router_defaults)